api now accepts images through posts and uploads them to cloudinary and returns a URL
